// src/app/models/system/policy/uniform.ts
export class UniformModel {
    constructor() {
             
    }
    
    uniform_id: string = "1";
    uniform_code: string = "";
    uniform_name_th: string = "";
    uniform_name_en: string = "";    
    modified_by: string = "";
    modified_date: string = "";
    index: number = 0;
    select: boolean = false;    
  }
  